requirements.txt
tensorflow==2.17.0
pandas==2.2.2
dvc==3.51.2
mlflow==2.2.2
notebook==7.0.6
numpy==1.26.4
matplotlib==3.8.0
seaborn==0.13.2
python-box==6.0.2
pyYAML==6.0.1
tqdm==4.66.5
ensure==1.0.2
joblib==1.4.2
types-pyYAML==6.0.12.20240311
scipy==1.14.1
flask==2.2.5
Flask-Cors==4.0.1
gdown==5.2.0
typing-extensions==4.12.2
ipython==8.18.1
pydantic==2.11.9
optree==0.17.0
Pillow
gradio
-e .